+++ hew-types/tests/trait_object_order.rs
@@ -6,10 +6,10 @@ use common::typecheck_isolated as typecheck;
 fn trait_object_different_order_unifies() {
     let output = typecheck(
         r"
-        trait A { fn a(val: Self) -> int; }
-        trait B { fn b(val: Self) -> int; }
-        fn takes_ab(x: dyn (A + B)) -> int { x.a() }
-        fn gives_ba(x: dyn (B + A)) -> int { takes_ab(x) }
+        trait A { fn a(val: Self) -> i64; }
+        trait B { fn b(val: Self) -> i64; }
+        fn takes_ab(x: dyn (A + B)) -> i64 { x.a() }
+        fn gives_ba(x: dyn (B + A)) -> i64 { takes_ab(x) }
     ",
     );
     assert!(
